# Docs - contract/openapi.yaml
openapi: 3.0.1
info:
  title: Homely API
  description: HTTP JSON dùng {statusCode,message,data}; DTO trong data dùng snake_case, VND là chuỗi số nguyên. Thanh toán và hoàn tiền là sandbox.
  version: 1.6.0
servers:
- description: Local backend; paths include /api/v1
  url: http://localhost:8080
paths:
...
  /api/v1/notifications/{id}/read:
    post:
      tags:
      - notification-controller
      summary: Mark notification as read
      operationId: markAsRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/NotificationDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/notifications/read-all:
    post:
      tags:
      - notification-controller
      summary: Mark all notifications as read
      operationId: markAllAsRead
      parameters:
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: integer
                      format: int32
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/notifications/mark-all-read:
    post:
      tags:
      - notification-controller
      summary: Mark all notifications as read
      operationId: markAllAsRead_1
      parameters:
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: integer
                      format: int32
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/media/upload:
    post:
      tags:
      - media-controller
      summary: Upload media file
      operationId: upload
      parameters:
      - name: purpose
        in: query
        required: true
        schema:
          type: string
          enum:
          - ROOM_PHOTO
          - ROOM_VIDEO
          - AVATAR
          - CHAT_IMAGE
          - REPORT_EVIDENCE
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/MediaDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/media/attach/room/{roomId}:
    post:
      tags:
      - media-controller
      summary: Attach media to room
      operationId: attachToRoom
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: purpose
        in: query
        required: false
        schema:
          type: string
          default: ROOM_PHOTO
          enum:
          - ROOM_PHOTO
          - ROOM_VIDEO
          - AVATAR
          - CHAT_IMAGE
          - REPORT_EVIDENCE
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachMediaRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/notifications:
    get:
      tags:
      - notification-controller
      summary: Get my notifications
      operationId: getMyNotifications
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 50
          minimum: 1
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PageResponseNotificationDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/notifications/unread-count:
    get:
      tags:
      - notification-controller
      summary: Get unread notification count
      operationId: getUnreadCount
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: integer
                      format: int64
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/media/{id}/content:
    get:
      tags:
      - media-controller
      summary: Read authorized media content
      operationId: content
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Raw file bytes; content type matches the stored file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/media/room/{roomId}:
    get:
      tags:
      - media-controller
      summary: Get room media
      operationId: getRoomMedia
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/media/my:
    get:
      tags:
      - media-controller
      summary: Get my ready media
      operationId: getMyReadyMedia
      parameters:
      - name: purpose
        in: query
        required: true
        schema:
          type: string
          enum:
          - ROOM_PHOTO
          - ROOM_VIDEO
          - AVATAR
          - CHAT_IMAGE
          - REPORT_EVIDENCE
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/media/{mediaId}:
    delete:
      tags:
      - media-controller
      summary: Delete media
      operationId: deleteMedia
      parameters:
      - name: mediaId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
components:
  schemas:
...
    NotificationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        title:
          type: string
        body:
          type: string
        ref_type:
          type: string
        ref_id:
          type: integer
          format: int64
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
    MediaDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        purpose:
          type: string
        status:
          type: string
        content_type:
          type: string
        file_size_bytes:
          type: integer
          format: int64
        original_filename:
          type: string
        duration_secs:
          type: integer
          format: int32
        display_order:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
    AttachMediaRequest:
      required:
      - mediaIds
      type: object
      properties:
        mediaIds:
          maxItems: 16
          minItems: 1
          type: array
          items:
            type: integer
            format: int64
...
    PageResponseNotificationDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/NotificationDTO'
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
        current_page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
...
    FieldErrorDTO:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ProblemDTO:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDTO'
        trace_id:
          type: string
    RestResponseProblemDTO:
      required:
      - statusCode
      - message
      - data
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
          example: 500
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProblemDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...